+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The changes documented here do not include those from the original repository.
 
 ## [Unreleased]
 
+## 2023-03-20
+- Feat: Update error code format
+- Feat: [Android] Update error codes and descriptions (https://outsystemsrd.atlassian.net/browse/RMET-2343)
+
 ### 2023-03-16
 - Feat: [iOS] Update error codes and descriptions (https://outsystemsrd.atlassian.net/browse/RMET-2342).